cleanup(core): consolidate core
