+++ packages/core/src/lib/di/performance.ts
