+++ packages/core/src/lib/pipes/color.pipe.ts
