+++ packages/core/src/lib/pipes/vector.pipe.ts
