+++ packages/core/src/lib/three/material.ts
@@ -12,7 +12,7 @@ import {
 import { requestAnimationFrame } from '@rx-angular/cdk/zone-less';
 import * as THREE from 'three';
 import { NgtContentMaterialController } from '../controllers/content-material.controller';
-import { NGT_OBJECT_3D } from '../providers/object3d';
+import { NGT_OBJECT_3D } from '../di/object3d';
 import { NgtStore } from '../stores/store';
 import type {
   AnyConstructor,
